storage: narrow assignment and worker ids to 32-bit (#96)
* storage: narrow assignment and worker ids to 32-bit

* docs: drop the sizing note and the worker-sync rationale from the schema reference

* migration: drop the width-rationale comments (rationale lives in docs/mvcc-schema.md)

* review: revert comment expansions; keep only the id-consumption test

* review: trim the worker-sync comments

* workers: note the NOT EXISTS screen relies on the single-scheduler lock

---------

Reviewed-on: http://localhost:3000/defnull/network-scheduler/pulls/96
Co-committed-by: claude <claude@example.com>

+++ src/multistep_scheduler/sim/sut.rs
@@ -529,7 +529,7 @@ impl<D: SimStorage> SimUnderTest<D> {
             .iter()
             .copied()
             .enumerate()
-            .map(|(i, worker)| (WorkerPk(i as i64 + 1), worker.clone()))
+            .map(|(i, worker)| (WorkerPk(i as i32 + 1), worker.clone()))
             .collect();
         let excluded = self.chunks_excluded_from_floor();
         let mut chunk_pairs: Vec<(ChunkPk, AlgoChunk)> = self
